Journal events to 3.3 beta 4
