+++ EliteDangerous/EliteDangerous/JournalEntry.cs
@@ -91,6 +91,7 @@ namespace EliteDangerousCore
         FetchRemoteModule = 1000,
         FSDJump = 280,
         FSDTarget = 281,
+        FSSAllBodiesFound = 285,
         FuelScoop = 290,
         Fileheader = 300,
         FighterDestroyed = 303,
@@ -142,6 +143,7 @@ namespace EliteDangerousCore
         ModuleSellRemote = 990,
         ModuleStore = 525,
         ModuleSwap = 530,
+        MultiSellExplorationData = 533,
         Music = 535,
         NavBeaconScan = 538,
         NewCommander = 540,
