+++ EliteDangerous/JournalEvents/JournalCargo.cs
@@ -13,11 +13,10 @@
  *
  * EDDiscovery is not affiliated with Frontier Developments plc.
  */
+
 using Newtonsoft.Json.Linq;
 using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 
 namespace EliteDangerousCore.JournalEvents
 {
@@ -46,8 +45,11 @@ namespace EliteDangerousCore.JournalEvents
         }
 
         void Rescan(JObject evt)
-        { 
+        {
+            Vessel = evt["Vessel"].Str("Ship");         // ship is the default.. present in 3.3 only.  Other value SRV
+
             Inventory = evt["Inventory"]?.ToObjectProtected<Cargo[]>().OrderBy(x => x.Name)?.ToArray();
+
             if (Inventory != null)
             {
                 foreach (Cargo c in Inventory)
@@ -76,12 +78,11 @@ namespace EliteDangerousCore.JournalEvents
             }
         }
 
-
+        public string Vessel { get; set; }          // always set, Ship or SRV.
         public Cargo[] Inventory { get; set; }      // may be NULL
 
         public override void FillInformation(out string info, out string detailed) 
         {
-            
             info = "No Cargo".Txb(this);
             detailed = "";
 
@@ -91,7 +92,9 @@ namespace EliteDangerousCore.JournalEvents
                 foreach (Cargo c in Inventory)
                     total += c.Count;
 
-                info = string.Format("Cargo, {0} items".Txb(this), total);
+                info = Vessel.Equals("Ship") ? "Ship ".Txb(this) : "SRV ".Txb(this);
+
+                info += string.Format( "Cargo, {0} items".Txb(this), total);
                 detailed = "";
 
                 foreach (Cargo c in Inventory)
@@ -108,13 +111,16 @@ namespace EliteDangerousCore.JournalEvents
 
         public void MaterialList(MaterialCommoditiesList mc, DB.SQLiteConnectionUser conn)
         {
-            //System.Diagnostics.Debug.WriteLine("Updated at " + this.EventTimeUTC.ToString());
-            mc.Clear(true);
-
-            if (Inventory != null)
+            if (Vessel.Equals("Ship"))      // only want ship cargo to change lists..
             {
-                foreach (Cargo c in Inventory)
-                    mc.Set(MaterialCommodityData.CommodityCategory, c.Name, c.Count, 0, conn);
+                //System.Diagnostics.Debug.WriteLine("Updated at " + this.EventTimeUTC.ToString());
+                mc.Clear(true);
+
+                if (Inventory != null)
+                {
+                    foreach (Cargo c in Inventory)
+                        mc.Set(MaterialCommodityData.CommodityCategory, c.Name, c.Count, 0, conn);
+                }
             }
         }
     }
@@ -163,5 +169,119 @@ namespace EliteDangerousCore.JournalEvents
             detailed = "";
         }
     }
+
+    [JournalEntryType(JournalTypeEnum.CargoDepot)]
+    public class JournalCargoDepot : JournalEntry, IMaterialCommodityJournalEntry
+    {
+        public JournalCargoDepot(JObject evt) : base(evt, JournalTypeEnum.CargoDepot)
+        {
+            MissionId = evt["MissionID"].Int();
+            UpdateType = evt["UpdateType"].Str();        // must be FD name
+            System.Enum.TryParse<UpdateTypeEnum>(UpdateType, out UpdateTypeEnum u);
+            UpdateEnum = u;
+            CargoType = evt["CargoType"].Str();     // item counts
+            FriendlyCargoType = MaterialCommodityData.GetNameByFDName(CargoType);
+            Count = evt["Count"].Int(0);
+            StartMarketID = evt["StartMarketID"].Long();
+            EndMarketID = evt["EndMarketID"].Long();
+            ItemsCollected = evt["ItemsCollected"].Int();
+            ItemsDelivered = evt["ItemsDelivered"].Int();
+            TotalItemsToDeliver = evt["TotalItemsToDeliver"].Int();
+            ItemsToGo = TotalItemsToDeliver - ItemsDelivered;
+            ProgressPercent = evt["Progress"].Double() * 100;
+            MarketID = evt["MarketID"].LongNull();
+
+            if (ProgressPercent < 0.01)
+                ProgressPercent = ((double)System.Math.Max(ItemsCollected, ItemsDelivered) / (double)TotalItemsToDeliver) * 100;
+        }
+
+        public enum UpdateTypeEnum { Unknown, Collect, Deliver, WingUpdate }
+
+        public int MissionId { get; set; }
+        public string UpdateType { get; set; }
+        public UpdateTypeEnum UpdateEnum { get; set; }
+
+        public string CargoType { get; set; } // 3.03       deliver/collect only    - what you have done now.  Blank if not known (<3.03)
+        public string FriendlyCargoType { get; set; }
+        public int Count { get; set; }  // 3.03         deliver/collect only.  0 if not known.
+
+        public long StartMarketID { get; set; }
+        public long EndMarketID { get; set; }
+
+        public int ItemsCollected { get; set; }             // current total stats
+        public int ItemsDelivered { get; set; }
+        public int ItemsToGo { get; set; }
+        public int TotalItemsToDeliver { get; set; }
+        public double ProgressPercent { get; set; }
+
+        public long? MarketID { get; set; }
+
+        public void MaterialList(MaterialCommoditiesList mc, DB.SQLiteConnectionUser conn)
+        {
+            if (CargoType.Length > 0 && Count > 0)
+                mc.Change(MaterialCommodityData.CommodityCategory, CargoType, (UpdateEnum == UpdateTypeEnum.Collect) ? Count : -Count, 0, conn);
+        }
+
+        public override void FillInformation(out string info, out string detailed)
+        {
+
+            if (UpdateEnum == UpdateTypeEnum.Collect)
+            {
+                info = BaseUtils.FieldBuilder.Build("Collected:".Txb(this), Count, "< of ".Txb(this), FriendlyCargoType, "Total:".Txb(this), ItemsDelivered, "To Go:", ItemsToGo, "Progress:;%;N1".Txb(this), ProgressPercent);
+            }
+            else if (UpdateEnum == UpdateTypeEnum.Deliver)
+            {
+                info = BaseUtils.FieldBuilder.Build("Delivered:".Txb(this), Count, "< of ".Txb(this), FriendlyCargoType, "Total:".Txb(this), ItemsDelivered, "To Go:", ItemsToGo, "Progress:;%;N1".Txb(this), ProgressPercent);
+            }
+            else if (UpdateEnum == UpdateTypeEnum.WingUpdate)
+            {
+                info = BaseUtils.FieldBuilder.Build("Update, Collected:".Txb(this), ItemsCollected, "Delivered:".Txb(this), ItemsDelivered, "To Go:".Txb(this), ItemsToGo, "Progress Left:;%;N1".Txb(this), ProgressPercent);
+            }
+            else
+            {
+                info = "Unknown CargoDepot type " + UpdateType;
+            }
+
+            detailed = "";
+        }
+    }
+
+
+    [JournalEntryType(JournalTypeEnum.CollectCargo)]
+    public class JournalCollectCargo : JournalEntry, IMaterialCommodityJournalEntry, ILedgerNoCashJournalEntry
+    {
+        public JournalCollectCargo(JObject evt) : base(evt, JournalTypeEnum.CollectCargo)
+        {
+            Type = evt["Type"].Str();                               //FDNAME
+            Type = JournalFieldNaming.FDNameTranslation(Type);     // pre-mangle to latest names, in case we are reading old journal records
+            FriendlyType = MaterialCommodityData.GetNameByFDName(Type);
+            Type_Localised = JournalFieldNaming.CheckLocalisationTranslation(evt["Type_Localised"].Str(), FriendlyType);         // always ensure we have one
+            Stolen = evt["Stolen"].Bool();
+            MissionID = evt["MissionID"].LongNull();
+        }
+
+        public string Type { get; set; }                    // FDNAME..
+        public string FriendlyType { get; set; }            // translated name
+        public string Type_Localised { get; set; }            // always set
+        public bool Stolen { get; set; }
+        public long? MissionID { get; set; }             // if applicable
+
+        public void MaterialList(MaterialCommoditiesList mc, DB.SQLiteConnectionUser conn)
+        {
+            mc.Change(MaterialCommodityData.CommodityCategory, Type, 1, 0, conn);
+        }
+
+        public void LedgerNC(Ledger mcl, DB.SQLiteConnectionUser conn)
+        {
+            mcl.AddEventNoCash(Id, EventTimeUTC, EventTypeID, FriendlyType);
+        }
+
+        public override void FillInformation(out string info, out string detailed)
+        {
+            info = BaseUtils.FieldBuilder.Build("", Type_Localised, ";Stolen".Txb(this), Stolen, "<; (Mission Cargo)".Txb(this), MissionID != null);
+            detailed = "";
+        }
+    }
+
 
 }
