+++ EliteDangerous/UIEvents/UICargo.cs
@@ -21,29 +21,15 @@ using System.Threading.Tasks;
 
 namespace EliteDangerousCore.UIEvents
 {
-    public class UIGUIFocus : UIEvent
+    public class UICargo : UIEvent
     {
-        public enum Focus
+        public UICargo(int value, UIShipType.Shiptype shiptype, DateTime time, bool refresh) : base(UITypeEnum.Cargo, time, refresh)
         {
-            NoFocus = 0,
-            SystemPanel = 1,
-            TargetPanel = 2,   
-            CommsPanel = 3, // top
-            RolePanel = 4,  // bottom
-            StationServices = 5,
-            GalaxyMap = 6,
-            SystemMap = 7,
-            Orrey=8,        //3.3
-            FSSMode =9, //3.3
-            SAAMode =10,//3.3
-            Codex =11,//3.3
+            Count = value;
+            ShipType = shiptype;
         }
 
-        public UIGUIFocus(int focus, DateTime time, bool refresh): base(UITypeEnum.GUIFocus, time, refresh)
-        {
-            GUIFocus = (Focus)focus;
-        }
-
-        public Focus GUIFocus { get; private set; }
+        public int Count { get; private set; }     // number of items
+        public UIShipType.Shiptype ShipType { get; private set; }   // Ship type flags.. per flags
     }
 }
