+++ EliteDangerous/UIEvents/UINightVision.cs
@@ -21,29 +21,13 @@ using System.Threading.Tasks;
 
 namespace EliteDangerousCore.UIEvents
 {
-    public class UIGUIFocus : UIEvent
+    public class UINightVision : UIEvent
     {
-        public enum Focus
+        public UINightVision(bool state, DateTime time, bool refresh): base(UITypeEnum.NightVision, time, refresh)
         {
-            NoFocus = 0,
-            SystemPanel = 1,
-            TargetPanel = 2,   
-            CommsPanel = 3, // top
-            RolePanel = 4,  // bottom
-            StationServices = 5,
-            GalaxyMap = 6,
-            SystemMap = 7,
-            Orrey=8,        //3.3
-            FSSMode =9, //3.3
-            SAAMode =10,//3.3
-            Codex =11,//3.3
+            State = state;
         }
 
-        public UIGUIFocus(int focus, DateTime time, bool refresh): base(UITypeEnum.GUIFocus, time, refresh)
-        {
-            GUIFocus = (Focus)focus;
-        }
-
-        public Focus GUIFocus { get; private set; }
+        public bool State { get; private set; }
     }
 }
